Generate Observed Data from choice model

Generates observed cognitive and choice data from the IRT-Thurstone model.

Generate_Choice(
  N,
  J,
  K,
  theta,
  as,
  bs,
  zeta,
  gamma,
  X,
  W,
  subject_ids,
  unique_subject_ids
)

Arguments

N

An integer for the number of observations.

J

An integer for the number of items.

K

An integer for the number of paired comparisons.

theta

A vector of latent cognitive variables.

as

A vector of length J with item discriminations.

bs

A vector of length J with item locations.

zeta

A matrix with dimensions N x V containing random parameter estimates.

gamma

A vector with dimensions P x 1 containing fixed parameter estimates, where \(P = P_1 + P_2\)

X

A matrix with dimensions N*K x P_1 containing fixed effect design matrix without theta.

W

A matrix with dimensions N*K x V containing random effect variables.

subject_ids

A vector with length NK x 1 containing subject-choice IDs.

unique_subject_ids

A vector with length N x 1 containing unique subject IDs.

Value

A list that contains:

Y

A matrix of dimension N by J

C

A vector of length NK

Author

Steven Andrew Culpepper and James Joseph Balamuta