Generate Random Multivariate Normal Distribution

Creates a random Multivariate Normal when given number of obs, mean, and sigma.

rmvnorm(n, mu, S)

Arguments

n
An integer, which gives the number of observations. (> 0)

mu
A vector length m that represents the means of the normals.

S
A matrix with dimensions m x m that provides Sigma, the covariance matrix.

Value

A matrix that is a Multivariate Normal distribution.

See also

Examples

# Call with the following data: 
rmvnorm(2, c(0,0), diag(2))

#>            [,1]       [,2]
#> [1,] -0.8342814 -0.3648162
#> [2,] -0.7915539 -0.1618449